requirements.txt
django
gunicorn
django-heroku
pandas
psycopg2-binary
openpyxl
